requirements.txt
google-cloud-storage>=2.19.0
google-cloud-aiplatform>=1.109.0
pillow>=9.4.0
pyyaml>=6.0
tenacity>=9.1.0
